+++ src/arkreactor/Compiler/Package/ImportSolver.cpp
@@ -1,4 +1,4 @@
-#include <Ark/Compiler/ImportSolver.hpp>
+#include <Ark/Compiler/Package/ImportSolver.hpp>
 
 #include <ranges>
 #include <algorithm>
